* Remove uses of MockBTAppContextSwitchDelegate from UnitTests * Rename MockDelegates.swift to MockViewControllerPresentingDelegate.swift
