+++ UnitTests/BraintreeTestShared/MockViewControllerPresentationDelegate.swift
@@ -1,0 +1,21 @@
+import XCTest
+import BraintreeCore
+
+@objc public class MockViewControllerPresentationDelegate : NSObject, BTViewControllerPresentingDelegate {
+    public var requestsPresentationOfViewControllerExpectation : XCTestExpectation? = nil
+    public var requestsDismissalOfViewControllerExpectation : XCTestExpectation? = nil
+    public var lastViewController : UIViewController? = nil
+    public var lastPaymentDriver : AnyObject? = nil
+
+    public func paymentDriver(_ driver: Any, requestsDismissalOf viewController: UIViewController) {
+        lastPaymentDriver = driver as AnyObject?
+        lastViewController = viewController
+        requestsDismissalOfViewControllerExpectation?.fulfill()
+    }
+
+    public func paymentDriver(_ driver: Any, requestsPresentationOf viewController: UIViewController) {
+        lastPaymentDriver = driver as AnyObject?
+        lastViewController = viewController
+        requestsPresentationOfViewControllerExpectation?.fulfill()
+    }
+}
